[BUG] rs232_init used uint8_t instead of uint16_t for the baud rate registers

+++ cpu/avr/dev/rs232.c
@@ -295,7 +295,7 @@ ISR(USART1_RX_vect)
 
 /*---------------------------------------------------------------------------*/
 void
-rs232_init(uint8_t port, uint8_t bd, uint8_t ffmt)
+rs232_init(uint8_t port, uint16_t bd, uint8_t ffmt)
 {
